back/requirements.txt
asgiref==3.2.7
astroid==2.3.3
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
Django==3.0.5
django-appconf==1.0.4
django-cors-headers==3.2.1
django-imagekit==4.0.2
django-rest-swagger==2.2.0
djangorestframework==3.11.0
dlib==19.19.0
drf-yasg==1.17.1
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==2.9
image==1.5.29
inflection==0.4.0
isort==4.3.21
itypes==1.1.0
Jinja2==2.11.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mysqlclient==1.4.6
mccabe==0.6.1
numpy==1.18.2
openapi-codec==1.3.2
packaging==20.3
pilkit==2.0
Pillow==7.1.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
pylint==2.4.4
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
wrapt==1.11.2
